src/tests/test_mail.py
from flask import Flask
from flask_mail import Mail, Message
from dotenv import load_dotenv
import os

load_dotenv()

app = Flask(__name__)

# Mail configuration
app.config['MAIL_SERVER'] = os.getenv('MAIL_SERVER')
app.config['MAIL_PORT'] = int(os.getenv('MAIL_PORT', 465))
app.config['MAIL_USE_TLS'] = os.getenv('MAIL_USE_TLS', 'False').lower() == 'true'
app.config['MAIL_USE_SSL'] = os.getenv('MAIL_USE_SSL', 'True').lower() == 'true'
app.config['MAIL_USERNAME'] = os.getenv('MAIL_USERNAME')
app.config['MAIL_PASSWORD'] = os.getenv('MAIL_PASSWORD')
app.config['MAIL_DEFAULT_SENDER'] = os.getenv('MAIL_DEFAULT_SENDER')

mail = Mail(app)

def test_email_config():
    print("\nTesting email configuration...")
    print(f"Mail Server: {app.config['MAIL_SERVER']}")
    print(f"Mail Port: {app.config['MAIL_PORT']}")
    print(f"Use TLS: {app.config['MAIL_USE_TLS']}")
    print(f"Use SSL: {app.config['MAIL_USE_SSL']}")
    print(f"Username: {app.config['MAIL_USERNAME']}")
    print(f"Password: {'*' * len(app.config['MAIL_PASSWORD']) if app.config['MAIL_PASSWORD'] else 'Not set'}")
    print(f"Default Sender: {app.config['MAIL_DEFAULT_SENDER']}")
    
    try:
        msg = Message(
            "Test Email from City Tagger",
            recipients=[app.config['MAIL_USERNAME']],
            body="This is a test email to verify the email configuration is working correctly."
        )
        mail.send(msg)
        print("\nTest email sent successfully!")
    except Exception as e:
        print(f"\nError sending test email: {str(e)}")

if __name__ == "__main__":
    with app.app_context():
        test_email_config() 