requirements.txt
requests==2.31.0
sqlalchemy==2.0.27
flask==3.0.2
python-dotenv==1.0.1
stripe==7.11.0
bcrypt==4.1.2
PyJWT==2.8.0
Flask-Mail==0.9.1
alembic==1.13.1
psycopg2-binary==2.9.9
Flask-Login==0.6.3
supabase==2.3.4
gunicorn==21.2.0 